+++ src/main/java/org/openpnp/machine/reference/ReferenceCamera.java
@@ -581,16 +581,13 @@ public abstract class ReferenceCamera extends AbstractBroadcastingCamera impleme
 
             if (advancedCalibration.isOverridingOldTransformsAndDistortionCorrectionSettings()) {
                 //Skip all the old style image transforms and distortion corrections except for 
-                //deinterlacing and cropping
+                //deinterlacing, cropping, and white balancing
                 if (isDeinterlaced() || isCropped() || isWhiteBalanced() || advancedCalibration.isEnabled()) {
                     Mat mat = OpenCvUtils.toMat(image);
                     mat = deinterlace(mat);
                     mat = crop(mat);
-                    if (advancedCalibration.isEnabled()) {
-                        //Use the new advanced image transformation and distortion correction
-                        mat = advancedUndistort(mat);
-                    }
                     mat = whiteBalance(mat);
+                    mat = advancedUndistort(mat);
                     image = OpenCvUtils.toBufferedImage(mat);
                     mat.release();
                 }
@@ -646,10 +643,11 @@ public abstract class ReferenceCamera extends AbstractBroadcastingCamera impleme
     }
 
     @Override
-    public Location getUnitsPerPixel(Length viewingPlaneZ) {
+    public synchronized Location getUnitsPerPixel(Length viewingPlaneZ) {
         if (advancedCalibration.isOverridingOldTransformsAndDistortionCorrectionSettings() && 
                 advancedCalibration.isValid()) {
-            double upp = advancedCalibration.getDistanceToCameraAtZ(viewingPlaneZ).
+            double upp = 0;
+            upp = advancedCalibration.getDistanceToCameraAtZ(viewingPlaneZ).
                         convertToUnits(LengthUnit.Millimeters).getValue() / 
                         advancedCalibration.getVirtualCameraMatrix().get(0, 0)[0];
             upp = Double.isFinite(upp) ? upp : 0;
@@ -658,15 +656,21 @@ public abstract class ReferenceCamera extends AbstractBroadcastingCamera impleme
         return super.getUnitsPerPixel(viewingPlaneZ);
     }
 
-    private Mat advancedUndistort(Mat mat) {
+    private synchronized Mat advancedUndistort(Mat mat) {
+        if (!advancedCalibration.isEnabled()) {
+            return mat;
+        }
+        Mat dst = mat.clone();
         if (undistortionMap1 == null || undistortionMap2 == null) {
-            undistortionMap1 = new Mat();
-            undistortionMap2 = new Mat();
+            if (undistortionMap1 == null) {
+                undistortionMap1 = new Mat();
+            }
+            if (undistortionMap2 == null) {
+                undistortionMap2 = new Mat();
+            }
             advancedCalibration.initUndistortRectifyMap(mat.size(), 
                     undistortionMap1, undistortionMap2);
         }
-        
-        Mat dst = mat.clone();
         Imgproc.remap(mat, dst, undistortionMap1, undistortionMap2, Imgproc.INTER_LINEAR);
         mat.release();
 
@@ -1111,6 +1115,12 @@ public abstract class ReferenceCamera extends AbstractBroadcastingCamera impleme
         return calibration.isEnabled();
     }
 
+    /**
+     * Flips the image. NOTE!!!!! - flipX means to flip about the x-axis which is a vertical flip
+     * and flipY means to flip about the y-axis which is a horizontal flip
+     * @param mat
+     * @return
+     */
     protected Mat flip(Mat mat) {
         if (isFlipped()) {
             int flipCode;